testing ownership visual cues
not fully implemented, but the idea is to be able to have a bunch of the
boxes like the one in Alaska, and there are 5 boxes, 1 for neutral, 1
for red player faction, 1 for green player faction... etc. If a player
owns a specific province, then the box changes color accordingly. If
possible (i dont know yet i havent looked that far into it) you could
also be able to mouse over the boxes and have it change into their
larger counterpart (see images folder for all the boxes)